More complex product import

+++ src/main/resources/db/migration/V1__Initial_shema.sql
@@ -3,7 +3,7 @@ CREATE SCHEMA mcr;
 CREATE TABLE products(
     id SERIAL PRIMARY KEY,
     title TEXT NOT NULL,
-    price DOUBLE PRECISION,
+    price NUMERIC(10,2),
     description TEXT,
     category TEXT
 );
